MM-26

+++ src/app/user/service.js
@@ -12,6 +12,7 @@ module.exports = {
     let { role, to, subject, text, html } = body;
     await sequelize.transaction(async (trx) => {
       const invite = await db.Invite.create({ to, role, restaurantId: restaurant.id }, { transaction: trx });
+      subject = `Work invitation in "${restaurant.name}"`;
       html = `You have been invited to join "${restaurant.name}" team as "${role}".<br /><a href="https://mymenusystem.herokuapp.com/invite?token=${invite.token}">Click here to accept.</a>`;
       await mailSender(to, subject, text, html);
     });
@@ -22,10 +23,7 @@ module.exports = {
     if (!invite) throw new Exception(statusCodes.ITEM_NOT_FOUND, "Invite token was not found.");
     body.password = await hash(body.password, Number(process.env.BCRYPT_ROUNDS));
     await sequelize.transaction(async (trx) => {
-      await Promise.all([
-        db.User.create({ ...body, restaurantId: invite.restaurantId, role: invite.role }, { transaction: trx }),
-        db.Invite.destroy({ where: { id: invite.id }, transaction: trx }),
-      ]);
+      await Promise.all([db.User.create({ ...body, restaurantId: invite.restaurantId, role: invite.role }, { transaction: trx }), db.Invite.destroy({ where: { id: invite.id }, transaction: trx })]);
     });
   },
 };
